// src/app/models/abstract/requestShipmentModelByProc.interface.ts
export interface RequestShipmentModelByProc{
    id: number;
    shipmentStatusId?: number;
    shipmentStatusName?: string;
    shipmentNumber?: string;
    isEnabled?: boolean;
    createdWhen?: any;
    orderDate?: any;
    fromHubId?: number;
    fromHubCode?: string;
    fromHubName?: string;
    pickUserId?: number;
    pickUserCode?: string;
    pickUserFullName?: string;
    senderId?: number;
    senderName?: string;
    senderPhone?: string;
    companyFrom?: string;
    pickingAddress?: string;
    fromProvinceId?: number;
    fromProvinceCode?: string;
    fromProvinceName?: string;
    fromHubRoutingId?: number;
    fromHubRoutingCode?: string;
    fromHubRoutingName?: string;
    receiverId?: number;
    receiverName?: string;
    receiverPhone?: string;
    companyTo?: string;
    shippingAddress?: string;
    toHubId?: number;
    toHubCode?: string;
    toHubName?: string;
    toProvinceId?: number;
    toProvinceCode?: string;
    toProvinceName?: string;
    toHubRoutingId?: number;
    toHubRoutingCode?: string;
    toHubRoutingName?: string;
    countShipmentAccept?: number;
    totalShipment?: number;
    totalShipmentNotFill?: number;
    cusNote?: string;
    isTooLate?: boolean;
}